New template

See merge request frenzz/ferramenta!13

+++ app/Http/Controllers/element_edit_controller.php
@@ -38,7 +38,7 @@ class element_edit_controller extends Controller
                 if(isset($request->file_name)){
                     //cancella foto
                     $path =  $element->pathPhoto;
-                    if(file_exists('storage'.$path)){
+                    if(file_exists('storage'.$path) && $path != null){
                         unlink(public_path('storage'.$path));
                     }
                     
@@ -88,6 +88,7 @@ class element_edit_controller extends Controller
                 if($courier->stima_giorni != $request->stima_giorni) $courier->update(['stima_giorni' => $request->stima_giorni]);
                 if($courier->price !== $request->price) $courier->update(['price' => $request->price]);
                 if($courier->name_service != $request->name_service) $courier->update(['name_service' => $request->name_service]);
+                if($courier->destination_country != $request->destination_country) $courier->update(['destination_country' => $request->destination_country]);
 
                 $path = $request-> ref;
                 $path = substr($path, 1, strlen($path));
